Game end screens
Added game over screen
Added victory screen
Moved all screens logic to one class
Added redirection to game over and victory screen in battle logic
