--- Assets/Scripts/OptionsScreen.cs
@@ -1,25 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class OptionsScreen : MonoBehaviour
-{
-    public string titleScreenScene; 
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void BackToTitleScreen()
-    {
-        SceneManager.LoadScene(titleScreenScene);
-    }
-}
