--- Assets/Scripts/TitleScreen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 64f8966e788ecd942877f068712bd626
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
