+++ ProjectSettings/EditorBuildSettings.asset
@@ -14,4 +14,10 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/OptionsScreen.unity
     guid: 8d517f9b82b248645a38a23736a55e20
+  - enabled: 1
+    path: Assets/Scenes/VictoryScreen.unity
+    guid: 659ff2f4d6f64d14a96ed36361d85c03
+  - enabled: 1
+    path: Assets/Scenes/GameOverScreen.unity
+    guid: 300dec80ee4fa5f498ba291220dce46a
   m_configObjects: {}
